refactor: 수량 선택 컴포넌트 분리(#48)
